Major Upgrades!
-Changed the delay on the sampling so that the nyquis frequency is now
abut 4500 Hz. It turned out that with a nyquist frequency of 20 kHz, all
the common frequencies in music were bunched up at the bottom of the
output, which made it difficult to actually use them. The highest note
on a piano is a bit under 4,200 Hz, so frequencies above that rarely
appear in actual music (at least based on my testing so far).
-Added outputs for LEDs for highs, mids, and lows